Extract cross-layer parse_spice_value to utils.format_utils
Move parse_spice_value() from simulation.monte_carlo to
utils.format_utils, eliminating the cross-layer dependency where
grading.circuit_comparer imported from the simulation layer.

Both monte_carlo.py (re-exports for backward compat) and
circuit_comparer.py now import from utils.format_utils.

Closes #771
