+++ app/utils/format_utils.py
@@ -193,6 +193,50 @@ def format_si(value, unit=""):
     return f"{scaled:.2f} {prefix}{unit}" if unit else f"{scaled:.2f} {prefix}"
 
 
+def parse_spice_value(value_str):
+    """Parse a SPICE value string (e.g. '1k', '100n', '4.7MEG') to float.
+
+    Unlike :func:`parse_value`, which raises ``ValueError`` on bad input, this
+    function returns ``None`` for unparseable strings (e.g. waveform specs like
+    ``'SIN(0 5 1k)'``).  This makes it suitable for contexts where a value
+    *might* be numeric but could also be a model name or function call.
+    """
+    s = value_str.strip()
+    if not s:
+        return None
+
+    # Map of SPICE suffix -> multiplier (ordered longest-first so "MEG" is
+    # tested before "M" etc.)
+    suffixes = [
+        ("MEG", 1e6),
+        ("meg", 1e6),
+        ("T", 1e12),
+        ("G", 1e9),
+        ("k", 1e3),
+        ("K", 1e3),
+        ("m", 1e-3),
+        ("u", 1e-6),
+        ("n", 1e-9),
+        ("p", 1e-12),
+        ("f", 1e-15),
+    ]
+
+    for suffix, mult in suffixes:
+        if s.endswith(suffix):
+            num_part = s[: -len(suffix)]
+            try:
+                return float(num_part) * mult
+            except ValueError:
+                return None
+
+    # Strip trailing unit letters (V, A, H, F, etc.)
+    stripped = s.rstrip("VAHFhvaf")
+    try:
+        return float(stripped)
+    except ValueError:
+        return None
+
+
 def format_frequency(freq_hz):
     """Format a frequency value with appropriate SI prefix."""
     if freq_hz is None:
